requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.5
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.13.0